small changes
revamp energy storage
recipe cancel logic for advanced sifter
config to set power capacity of machines
move base power consumption to recipe.json
prep for mining rig gui

+++ src/main/java/net/milkev/milkevsoreminers/common/blockEntities/miningRig/MiningRigTier2BlockEntity.java
@@ -48,12 +48,12 @@ public class MiningRigTier2BlockEntity extends MiningRigBaseBlockEntity {
     }
 
     @Override
-    public int getPowerCost() {
+    public long getPowerCost() {
         return 1;/*this.recipe.getPowerCost();*/
     }
 
     @Override
-    public int getPowerUsageSpeed() {
+    public long getPowerUsageSpeed() {
         return 0;
     }
